+++ termlink/gsea.py
@@ -5,18 +5,19 @@ defined by the Geneset dataset.
 
 The download files for Geneset are provided at http://software.broadinstitute.org/gsea/msigdb/collections.jsp.
 """
-import os
+
 import csv
 import json
 
+from os import path
+from re import match
 from urllib.parse import urlparse
 
-from termlink.commands import SubCommand
 from termlink.models import Coding, Relationship, RelationshipSchema
-from termlink.services import RelationshipService
 
+_filename_regex = r'msigdb\..*\.symbols\.gmt'
 
-def _to_json(rec, index):
+def _to_relationship(rec, index):
     """
     Convert record in table to Relationship as a JSON object
 
@@ -29,7 +30,6 @@ def _to_json(rec, index):
     Returns:
         A new record containing a single field, which is the JSON object
     """
-
     source = Coding(
         system="http://www.broadinstitute.org/gsea/msigdb",
         code=rec[index],
@@ -42,77 +42,41 @@ def _to_json(rec, index):
         display=rec[0]
     )
 
-    relationship = Relationship('subsumes', source, target)
-
-    schema = RelationshipSchema()
-    return json.dumps(schema.dump(relationship), sort_keys=True)
+    return Relationship('subsumes', source, target)
 
 
-class Command(SubCommand):
-    """
-    A command executor for GSEA operations
-    """
-
-    @staticmethod
-    def execute(args):
-        """
-        Prints a JSON array of `Relationship` objects to stdout
-
-        Args:
-            args: `argparse` parsed arguments
-        """
-        uri = urlparse(args.uri)
-        service = Service(uri)
-        rows = service.get_relationships()
-        for row in rows:
-            print(row)
-
-
-def validate_path(path):
-    """
-     Validates the file name for GSEA
+def _get_relationships(uri):
+    '''Extracts the system entities from the GSEA file
 
     Args:
-        path: The path/filename of the GSEA file
-    """
-    file_splits = os.path.basename(path).split('.')
-    if file_splits[0] != 'msigdb':
-        raise ValueError("only 'msigdb' is supported, %s not supported" % file_splits[0])
+        uri: a URI for the GSEA file on the local filesystem
 
-    if file_splits[3] != 'symbols':
-        raise ValueError("only 'symbols' is supported, %s not supported" % file_splits[3])
+    Returns:
+        yields relationships
+    '''
+    with open(uri.path) as f:
+        reader = csv.reader(f, delimiter='\t')
+        for row in reader:
+            for i in range(2, len(row)):
+                yield _to_relationship(row, i)
 
-    if file_splits[4] != 'gmt':
-        raise ValueError("file type %s not supported" % file_splits[4])
 
+def execute(args):
+    '''Converts the GSEA ontology.
 
-class Service(RelationshipService):
-    """Converts the GSEA database"""
+    Args:
+        args:   command line arguments from argparse
+    '''
+    uri = urlparse(args.uri)
+    if uri.scheme != 'file':
+        raise ValueError(f"uri.scheme '{uri.scheme}' is not supported")
 
-    def __init__(self, uri):
-        """
-        Bootstraps a service
+    filename = path.basename(uri.path) 
+    if not match(_filename_regex, filename):
+        raise ValueError(
+            f"File type is incorrect. Expected to match regular expression: '{_filename_regex}'. Found '{filename}'.")
 
-        Args:
-            uri: URI to root location of .gmt files
-        """
-
-        if uri.scheme != 'file':
-            raise ValueError("'uri.scheme' %s not supported" % uri.scheme)
-
-        self.uri = uri
-
-    def get_relationships(self):
-        """
-         Extracts the system entities from the GSEA file
-
-        Returns:
-            A new record containing a system, which is the JSON object
-        """
-        validate_path(self.uri.path)
-
-        with open(self.uri.path) as gsea_tsv:
-            gsea_reader = csv.reader(gsea_tsv, delimiter='\t')
-            for gsea in gsea_reader:
-                for i in range(2, len(gsea)):
-                    yield _to_json(gsea, i)
+    schema = RelationshipSchema()
+    for relationship in _get_relationships(uri):
+        o = json.dumps(schema.dump(relationship)) 
+        print(o)