+++ tests/client_test.py
@@ -1,6 +1,6 @@
 """Validates the 'client.py" module"""
 
-from nose.tools import ok_, eq_
+from nose.tools import ok_, eq_, raises
 
 from requests_mock import Adapter
 
@@ -9,13 +9,14 @@ from termlink.client import Client
 
 def test_client_construction():
     """Checks that the client can be instantiated"""
-
     url = 'url'
-
     client = Client(url)
-
     ok_(client)
 
+@raises(TypeError)
+def test_url_is_required():
+    """Checks that the url parameter is required"""
+    client = Client(url=None)
 
 def test_mocking_client_adapter():
     """Checks that mocking the client adapter is possible"""