+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Alters the output format for the `common` sub-command to JSON Lines
+- Improves the runtime performance of the `gsea` sub-command
+- Reduces the memory footprint of the `gsea` sub-command
+
 ## 0.8.1
 
 ### Fixed
@@ -44,7 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds support for .obo file types to `common` sub-command
 - Adds support for .owl file types to `common` sub-command
 
-### Changes
+### Changed
 
 - Improved performance of internal JSON serialization.